creating files to modularize the code in the breakdown portion of the component, bug in ProductBreakdown will be worked on in future
