+++ client/src/components/reviews/breakdownComponents/AverageRating.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+let AverageRating = () => {
+  return(<div>
+    <p>3.6</p>
+    <p>ALL OF THE STARS</p>
+    <p>Total Number of Reviews</p>
+  </div>);
+}
+
+export {AverageRating};
