+++ client/src/components/reviews/breakdownComponents/ProductBreakdown.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+let ProductBreakdown = () => {
+  return (<div>
+    <p>Qualities rating graph</p>
+  </div>);
+}
