+++ client/src/components/reviews/breakdownComponents/RatingBreakdown.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+let RatingBreakdown = () => {
+  return (<div>
+    <p><a>5 Stars</a> Graph </p>
+    <p><a>5 Stars</a> Graph </p>
+  </div>);
+}
+
+export {RatingBreakdown};
